--- src/nativeapm/src/components/StatusIcon/index.tsx
@@ -1,5 +1,0 @@
-const StatusIcon = () => {
-  return <button>StatusIcon</button>;
-};
-
-export default StatusIcon;
