+++ src/nativeapm/src/modules/api-data/mappers/gateway.ts
