+++ src/nativeapm/src/providers/AppProviders.tsx
@@ -1,0 +1,14 @@
+import { ThemeProvider } from '../modules/widget-theme';
+import { PaymentDataProvider } from '../modules/api-data';
+
+type AppProvidersPropsType = {
+  children: React.ReactElement;
+};
+
+const AppProviders = ({ children }: AppProvidersPropsType) => (
+  <PaymentDataProvider>
+    <ThemeProvider>{children}</ThemeProvider>
+  </PaymentDataProvider>
+);
+
+export default AppProviders;
