+++ src/nativeapm/src/providers/index.tsx
@@ -1,14 +1,1 @@
-import { ThemeProvider } from '../modules/theme';
-import { PaymentDataProvider } from '../modules/payments';
-
-type AppProvidersPropsType = {
-  children: React.ReactElement;
-};
-
-const AppProviders = ({ children }: AppProvidersPropsType) => (
-  <PaymentDataProvider>
-    <ThemeProvider>{children}</ThemeProvider>
-  </PaymentDataProvider>
-);
-
-export default AppProviders;
+export { default as AppProviders } from './AppProviders';
